backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
httpx>=0.27.0
python-dotenv>=1.0.1
biopython>=1.84
numpy>=1.26.4
scikit-learn>=1.5.1
joblib>=1.4.2
pydantic>=2.8.2
pydantic-settings>=2.4.0
pandas>=2.2.2
python-multipart>=0.0.9
